+++ website/i18n/en.json
@@ -13,7 +13,6 @@
     "drone-api": "Drone API Overview",
     "getting-started": "Getting Started",
     "Getting Started": "Getting Started",
-    "how-it-works": "How It Works",
     "old-readme": "old-readme",
     "visdom-tutorial": "Plotting Streams Of Data",
     "Plotting Streams Of Data": "Plotting Streams Of Data",
