+++ website/sidebars.json
@@ -15,8 +15,7 @@
       "visdom-tutorial"
     ],
     "Development": [
-      "dev-setup",
-      "how-it-works"
+      "dev-setup"
     ]
   }
 }
